Revert to v2.0.0
This reverts most of the changes made in v3 & v4. Mainly, this gets rid of the custom test runner and relies on xunit's built-in test-running logic. This will eliminate flakiness in some of our builds.

Also removing the ISpecification interface and instead relying on AsyncSpecification & Specification base classes.

Updating documentation to reflect the new state of things.

+semver:major